+++ src/test/java/com/yapp/artie/domain/user/application/service/GetUserServiceTest.java
@@ -8,6 +8,7 @@ import static org.mockito.BDDMockito.given;
 
 import com.yapp.artie.domain.user.application.port.out.LoadUserPort;
 import com.yapp.artie.domain.user.domain.User;
+import com.yapp.artie.domain.user.domain.UserNotFoundException;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
