+++ src/test/java/com/yapp/artie/domain/user/application/service/GetUserThumbnailServiceTest.java
@@ -10,6 +10,7 @@ import com.yapp.artie.domain.archive.service.ExhibitService;
 import com.yapp.artie.domain.user.application.port.in.GetUserThumbnailResponse;
 import com.yapp.artie.domain.user.application.port.out.LoadUserPort;
 import com.yapp.artie.domain.user.domain.User;
+import com.yapp.artie.domain.user.domain.UserNotFoundException;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
