+++ src/test/java/com/yapp/artie/domain/user/application/service/RenameUserServiceTest.java
@@ -9,6 +9,7 @@ import static org.mockito.BDDMockito.given;
 import com.yapp.artie.domain.user.application.port.out.LoadUserPort;
 import com.yapp.artie.domain.user.application.port.out.UpdateUserStatePort;
 import com.yapp.artie.domain.user.domain.User;
+import com.yapp.artie.domain.user.domain.UserNotFoundException;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
